feat(.env.exemple): Added port, database uri and node environement.
feat(src/config/db.js): Configured the connection to the database and handling failed attemtps when connecting to the database.

feat(src/app.js):  Configured the server and added an health route to check if everything is working good.

feat(src/server.js): Established connection to the database after launched the server.