feat(app.js): Added task routes.
feat(models/task.js): Added task model and added a status.

feat(task/taskcontroller.js): Added a CRUD with a try catch section for every methods.

feat(task/taskRoutes.js): Added a CRUD routes and a middlware to check the validity of the token for each of them.

feat(middlewares/authMiddleware.js):  Get the token, Check if the token is valid, check if the token exists, Add user to the request and handle errors.